🎨 Split the code into separate files
