+++ code/__init__.py
@@ -1,0 +1,3 @@
+from .github_flow import GitHubFlow
+from .github_pr import GitHubPR
+from .github_users import GitHubUsers
